chore: remove non-tls integration test for show queries multi node
