+++ ksqldb-rest-app/src/test/java/io/confluent/ksql/rest/integration/ShowQueriesMultiNodeWithTlsFunctionalTest.java
@@ -16,20 +16,27 @@
 package io.confluent.ksql.rest.integration;
 
 import static io.confluent.ksql.test.util.AssertEventually.assertThatEventually;
+import static org.hamcrest.Matchers.containsInAnyOrder;
 import static org.hamcrest.Matchers.is;
 
 import io.confluent.common.utils.IntegrationTest;
 import io.confluent.ksql.integration.IntegrationTestHarness;
 import io.confluent.ksql.integration.Retry;
 import io.confluent.ksql.rest.entity.KsqlEntity;
+import io.confluent.ksql.rest.entity.KsqlHostInfoEntity;
 import io.confluent.ksql.rest.entity.Queries;
+import io.confluent.ksql.rest.entity.QueryDescription;
+import io.confluent.ksql.rest.entity.QueryDescriptionList;
 import io.confluent.ksql.rest.entity.RunningQuery;
 import io.confluent.ksql.rest.server.KsqlRestConfig;
 import io.confluent.ksql.rest.server.TestKsqlRestApp;
 import io.confluent.ksql.serde.FormatFactory;
 import io.confluent.ksql.test.util.secure.ServerKeyStore;
 import io.confluent.ksql.util.PageViewDataProvider;
+
+import java.util.Collections;
 import java.util.List;
+import java.util.Set;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Supplier;
 import kafka.zookeeper.ZooKeeperClientException;
@@ -46,6 +53,8 @@ public class ShowQueriesMultiNodeWithTlsFunctionalTest {
   private static final String PAGE_VIEW_TOPIC = PAGE_VIEWS_PROVIDER.topicName();
   private static final String PAGE_VIEW_STREAM = PAGE_VIEWS_PROVIDER.kstreamName();
   private static final IntegrationTestHarness TEST_HARNESS = IntegrationTestHarness.build();
+  private static final KsqlHostInfoEntity host0 = new KsqlHostInfoEntity("localhost", 8089);
+  private static final KsqlHostInfoEntity host1 = new KsqlHostInfoEntity("localhost", 8099);
   private static final TestKsqlRestApp REST_APP_0 = TestKsqlRestApp
       .builder(TEST_HARNESS::kafkaBootstrapServers)
       .withProperty(KsqlRestConfig.LISTENERS_CONFIG,
@@ -118,4 +127,40 @@ public class ShowQueriesMultiNodeWithTlsFunctionalTest {
     return runningQueries.get(0).getStatusCount().toString();
   }
 
+  @Test
+  public void shouldShowAllQueriesExtended() {
+    // When:
+    final Supplier<Set<KsqlHostInfoEntity>> app0Response = () -> getShowQueriesExtendedResult(REST_APP_0);
+    final Supplier<Set<KsqlHostInfoEntity>> app1Response = () -> getShowQueriesExtendedResult(REST_APP_1);
+
+    // Then:
+    assertThatEventually("App0", app0Response, containsInAnyOrder(host0, host1));
+    assertThatEventually("App1", app1Response, containsInAnyOrder(host0, host1));
+  }
+
+  private static Set<KsqlHostInfoEntity> getShowQueriesExtendedResult(final TestKsqlRestApp restApp) {
+    final List<KsqlEntity> results = RestIntegrationTestUtil.makeKsqlRequest(
+        restApp,
+        "Show Queries Extended;"
+    );
+
+    if (results.size() != 1) {
+      return Collections.emptySet();
+    }
+
+    final KsqlEntity result = results.get(0);
+
+    if (!(result instanceof QueryDescriptionList)) {
+      return Collections.emptySet();
+    }
+
+    final List<QueryDescription> queryDescriptions = ((QueryDescriptionList) result)
+        .getQueryDescriptions();
+
+    if (queryDescriptions.size() != 1) {
+      return Collections.emptySet();
+    }
+
+    return queryDescriptions.get(0).getKsqlHostQueryStatus().keySet();
+  }
 }
